Move credentials resolving from CLI to SDK
