+++ cli/build.gradle
@@ -39,12 +39,10 @@ version = getenv('TRAVIS_TAG') ? getenv('TRAVIS_TAG') : "0.0.1";
 dependencies {
     compile project(':sdk')
 
-    compile "com.amazonaws:aws-java-sdk-iam:$awsVersion"
     compile "com.amazonaws:aws-java-sdk-sts:$awsVersion"
     compile "com.amazonaws:aws-java-sdk-core:$awsVersion"
 
     compile "com.fasterxml.jackson.core:jackson-core:$jacksonVersion"
-    compile "com.fasterxml.jackson.core:jackson-annotations:$jacksonVersion"
     compile "com.fasterxml.jackson.core:jackson-databind:$jacksonVersion"
     compile "com.fasterxml.jackson.datatype:jackson-datatype-jdk8:$jacksonVersion"
     compile "com.google.guava:guava:$guavaVersion"
